backend-deploy/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
motor==3.3.2
bson==0.5.10
cloudinary==1.38.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.2.0
opencv-python-headless==4.9.0.80
loguru==0.7.2
mangum==0.17.0
# Celery worker dependencies
celery[redis]==5.3.4
pymongo==4.6.1
numpy==1.26.3
torch>=2.2.0
diffusers==0.26.3
transformers==4.37.2
accelerate==0.26.1

